Fix XrmResource layout if pointers are bigger than long
On CHERI-enabled architectures (e.g. Arm's Morello), pointers are twice
the size of addresses (i.e. 128 bits for Morello, 64 bits for 32-bit
RISC-V). However, XtArgVal is currently defined as long, so it cannot
be used to store pointers on these architectures.

Also add a _Static_assert() when compiling with C11 support to check
that the offset of the last member matches that of XtResource.

Signed-off-by: Alex Richardson <Alexander.Richardson@cl.cam.ac.uk>